requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-SocketIO
python-dotenv
